# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionStatusStatus = typeof SessionStatusStatus[keyof typeof SessionStatusStatus];


export const SessionStatusStatus = {
  disconnected: 'disconnected',
  connecting: 'connecting',
  connected: 'connected',
  error: 'error',
} as const;

export interface SessionStatus {
  connected: boolean;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  lastSeen?: string | null;
  status: SessionStatusStatus;
}

export interface PairingCode {
  code: string;
  expiresAt: string;
  instructions: string;
}

export interface OperationResult {
  success: boolean;
  message: string;
}

export type MessageDirection = typeof MessageDirection[keyof typeof MessageDirection];


export const MessageDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  received: 'received',
  sent: 'sent',
  delivered: 'delivered',
  read: 'read',
} as const;

export interface Message {
  id: string;
  contactId: string;
  contactName: string;
  text: string;
  direction: MessageDirection;
  timestamp: string;
  isAutoReply: boolean;
  status: MessageStatus;
}

export interface Contact {
  id: string;
  name: string;
  phoneNumber: string;
  lastMessageAt: string;
  messageCount: number;
  unreadCount: number;
}

export interface ReplyInput {
  text: string;
}

export interface BotConfig {
  enabled: boolean;
  businessName: string;
  /** @nullable */
  gstNumber?: string | null;
  greeting: string;
  refundPolicy: string;
  autoReplyDelay: number;
  workingHoursEnabled: boolean;
  workingHoursStart: string;
  workingHoursEnd: string;
}

export interface BotConfigInput {
  enabled?: boolean;
  businessName?: string;
  /** @nullable */
  gstNumber?: string | null;
  greeting?: string;
  refundPolicy?: string;
  autoReplyDelay?: number;
  workingHoursEnabled?: boolean;
  workingHoursStart?: string;
  workingHoursEnd?: string;
}

export type BotRuleMatchType = typeof BotRuleMatchType[keyof typeof BotRuleMatchType];


export const BotRuleMatchType = {
  exact: 'exact',
  contains: 'contains',
  startsWith: 'startsWith',
} as const;

export interface BotRule {
  id: number;
  keyword: string;
  response: string;
  enabled: boolean;
  matchType: BotRuleMatchType;
  triggerCount: number;
}

export type BotRuleInputMatchType = typeof BotRuleInputMatchType[keyof typeof BotRuleInputMatchType];


export const BotRuleInputMatchType = {
  exact: 'exact',
  contains: 'contains',
  startsWith: 'startsWith',
} as const;

export interface BotRuleInput {
  keyword: string;
  response: string;
  enabled?: boolean;
  matchType: BotRuleInputMatchType;
}

export interface StatsSummary {
  totalMessages: number;
  inboundToday: number;
  autoRepliedToday: number;
  activeContacts: number;
  botEnabled: boolean;
  sessionConnected: boolean;
  responseRate: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  message_received: 'message_received',
  auto_reply_sent: 'auto_reply_sent',
  manual_reply_sent: 'manual_reply_sent',
  session_connected: 'session_connected',
  session_disconnected: 'session_disconnected',
  rule_triggered: 'rule_triggered',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  contactName?: string | null;
  timestamp: string;
}

export type GetMessagesParams = {
limit?: number;
/**
 * @nullable
 */
contactId?: string | null;
};