# lib/api-zod/src/generated/types/botConfigInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfigInput {
  enabled?: boolean;
  businessName?: string;
  /** @nullable */
  gstNumber?: string | null;
  greeting?: string;
  refundPolicy?: string;
  autoReplyDelay?: number;
  workingHoursEnabled?: boolean;
  workingHoursStart?: string;
  workingHoursEnd?: string;
}