# lib/api-zod/src/generated/types/botRule.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotRuleMatchType } from './botRuleMatchType';

export interface BotRule {
  id: number;
  keyword: string;
  response: string;
  enabled: boolean;
  matchType: BotRuleMatchType;
  triggerCount: number;
}