# lib/api-zod/src/generated/types/botRuleInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotRuleInputMatchType } from './botRuleInputMatchType';

export interface BotRuleInput {
  keyword: string;
  response: string;
  enabled?: boolean;
  matchType: BotRuleInputMatchType;
}