# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: string;
  name: string;
  phoneNumber: string;
  lastMessageAt: string;
  messageCount: number;
  unreadCount: number;
}