# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageDirection } from './messageDirection';
import type { MessageStatus } from './messageStatus';

export interface Message {
  id: string;
  contactId: string;
  contactName: string;
  text: string;
  direction: MessageDirection;
  timestamp: string;
  isAutoReply: boolean;
  status: MessageStatus;
}