# lib/api-zod/src/generated/types/pairingCode.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PairingCode {
  code: string;
  expiresAt: string;
  instructions: string;
}