# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalMessages: number;
  inboundToday: number;
  autoRepliedToday: number;
  activeContacts: number;
  botEnabled: boolean;
  sessionConnected: boolean;
  responseRate: number;
}